Address L Cluster "thin-only" issue
Shrink should not break connectivity, but the test provided will do so.
Perhaps more surprisingly, this also appears to be a bug in the
GNU Octave implementation of shrink as well. See Bug #44396
https://savannah.gnu.org/bugs/?44396

I was working from Digital Image Processing, 3rd Edition, by
William K. Pratt. It appears that the label stating that L Clusters
are "thin-only" was removed in the 4th Edition. This change fixes the
connectivity, though the result is still not the same as Matlab.

Various implementations of binary image shrink, as of today:

Input:
 [0, 0, 0, 1, 0;
  1, 1, 1, 1, 0;
  0, 0, 1, 1, 0;
  0, 0, 1, 1, 0;
  0, 0, 0, 1, 0]

My Shrink:
 [0, 0, 0, 0, 0;
  0, 1, 1, 1, 0;
  0, 0, 0, 1, 0;
  0, 0, 0, 1, 0;
  0, 0, 0, 1, 0]

My Shrink Max:
 [0, 0, 0, 0, 0;
  0, 0, 0, 1, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 0, 0]

Matlab Shrink:
 [0, 0, 0, 0, 0;
  0, 1, 1, 0, 0;
  0, 0, 0, 1, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 0, 0]

Matlab Shrink Max:
 [0, 0, 0, 0, 0;
  0, 0, 1, 0, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 0, 0]

Octave Shrink:
 [0, 0, 0, 0, 0;
  0, 1, 1, 1, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 1, 0;
  0, 0, 0, 1, 0]

Octave Shrink Max
 [0, 0, 0, 0, 0;
  0, 0, 1, 0, 0;
  0, 0, 0, 0, 0;
  0, 0, 0, 1, 0;
  0, 0, 0, 0, 0]